Make biopython optional
