+++ pyproject.toml
@@ -49,12 +49,13 @@ ci = ["pytest-github-actions-annotate-failures"]
 test-scripts = []
 test-scripts-all = ["geomstats[test-scripts]", "scikeras", "tensorflow"]
 graph = ["networkx"]
+bhv = ["networkx", "biopython", "pygraphviz"]
 pykeops = ["pykeops"]
 autograd = ["autograd >= 1.3"]
 pytorch = ["torch >= 1.9.1"]
 backends = ["geomstats[autograd, pytorch]"]
 optim-pytorch = ["geomstats[pytorch]", "pytorch-minimize"]
-opt = ["geomstats[graph, pykeops, optim-pytorch]"]
+opt = ["geomstats[graph, bhv, pykeops, optim-pytorch]"]
 dev = ["geomstats[test, test-scripts, lint, doc]"]
 all = ["geomstats[dev, opt, backends]"]
